+++ packages/fluentui/react-northstar/src/themes/teams-high-contrast/colors.ts
@@ -3,7 +3,7 @@ import { TeamsCategoryColorSchemeMapping } from '../teams/types';
 import { ColorSchemeMapping } from '../../themes/types';
 import { createColorScheme as createEmptyColorScheme } from '../teams/colors';
 
-export const accessibleYellow = '#ffff01';
+export const accessibleYellow = '#ffff00';
 export const accessibleGreen = '#3ff23f'; // always disabled color in high contrast
 export const accessibleCyan = '#1aebff';
 export const red = '#f00';
@@ -85,7 +85,7 @@ export const colorScheme: ColorSchemeMapping = {
     backgroundHover1: accessibleCyan,
     backgroundHover2: accessibleCyan,
     backgroundHover3: colors.black,
-    backgroundHover4: colors.black,
+    backgroundHover4: accessibleCyan,
 
     borderHover: accessibleCyan,