+++ packages/react-spinbutton/src/stories/SpinButton.stories.tsx
@@ -14,7 +14,8 @@ export { Size } from './SpinButtonSize.stories';
 export { Appearance } from './SpinButtonAppearance.stories';
 export { RTL } from './SpinButtonRTL.stories';
 export { Disabled } from './SpinButtonDisabled.stories';
-export { InputType } from './SpinButtonInputType.stories';
+export { Strings } from './SpinButtonStrings.stories';
+
 import { makeStyles, mergeClasses, shorthands } from '@griffel/react';
 
 const useDecoratorStyles = makeStyles({
@@ -32,7 +33,7 @@ const useDecoratorStyles = makeStyles({
 });
 
 const meta: Meta = {
-  title: 'Components/SpinButton',
+  title: 'Preview Components/SpinButton',
   component: SpinButton,
   parameters: {
     docs: {