+++ packages/react-theme/src/index.ts
@@ -3,6 +3,7 @@ export * from './utils/index';
 
 export { themeToTokensObject } from './themeToTokensObject';
 export { tokens } from './tokens';
+export { typographyStyles } from './global/index';
 
 export type {
   Brands,
@@ -12,6 +13,11 @@ export type {
   LineHeightTokens,
   BorderRadiusTokens,
   StrokeWidthTokens,
+  SpacingTokens,
+  HorizontalSpacingTokens,
+  VerticalSpacingTokens,
+  DurationTokens,
+  CurveTokens,
   ShadowTokens,
   ShadowBrandTokens,
   FontFamilyTokens,
@@ -68,4 +74,5 @@ export type {
   ColorTokens,
   PartialTheme,
   Theme,
+  TypographyStyles,
 } from './types';