+++ packages/react-theme/src/utils/createHighContrastTheme.ts
@@ -1,7 +1,12 @@
-import { colorPaletteTokens, generateColorTokens } from '../alias/highContrast';
+import { colorPaletteTokens } from '../alias/highContrastColorPalette';
+import { generateColorTokens } from '../alias/highContrastColor';
+
 import { borderRadius, fontSizes, lineHeights, fontFamilies, strokeWidths, fontWeights } from '../global/index';
 import { createShadowTokens } from './shadows';
 import type { Theme } from '../types';
+import { durations } from '../global/durations';
+import { curves } from '../global/curves';
+import { horizontalSpacings, verticalSpacings } from '../global/spacings';
 
 export const createHighContrastTheme = (): Theme => {
   const colorTokens = generateColorTokens();
@@ -13,6 +18,10 @@ export const createHighContrastTheme = (): Theme => {
     ...fontFamilies,
     ...fontWeights,
     ...strokeWidths,
+    ...horizontalSpacings,
+    ...verticalSpacings,
+    ...durations,
+    ...curves,
 
     ...colorTokens,
     ...colorPaletteTokens,