+++ packages/react-theme/src/utils/createLightTheme.ts
@@ -1,7 +1,12 @@
-import { colorPaletteTokens, generateColorTokens } from '../alias/light';
+import { colorPaletteTokens } from '../alias/lightColorPalette';
+import { generateColorTokens } from '../alias/lightColor';
+
 import { borderRadius, fontSizes, lineHeights, fontFamilies, strokeWidths, fontWeights } from '../global/index';
 import { createShadowTokens } from './shadows';
 import type { BrandVariants, Theme } from '../types';
+import { durations } from '../global/durations';
+import { curves } from '../global/curves';
+import { horizontalSpacings, verticalSpacings } from '../global/spacings';
 
 export const createLightTheme: (brand: BrandVariants) => Theme = brand => {
   const colorTokens = generateColorTokens(brand);
@@ -13,6 +18,10 @@ export const createLightTheme: (brand: BrandVariants) => Theme = brand => {
     ...fontFamilies,
     ...fontWeights,
     ...strokeWidths,
+    ...horizontalSpacings,
+    ...verticalSpacings,
+    ...durations,
+    ...curves,
 
     ...colorTokens,
     ...colorPaletteTokens,